remove sites faciles feature flipping #3205
