+++ .env.sample
@@ -60,7 +60,6 @@ ALBERT_API_KEY=
 SEND_ERROR_EMAIL_TOKEN=
 SIGNAL_LOGEMENT_PROD_URL=
 SITES_FACILES_URL=
-FEATURE_SITES_FACILES=0
 PLATFORM_NAME='Signal Logement'
 PLATFORM_LOGO='logo-signal-logement.svg'
 FEATURE_SECURE_UUID_URL=1
