+++ config/services.yaml
@@ -49,7 +49,6 @@ parameters:
     standard_max_list_pagination: 50
     feature_bo_signalement_create: '%env(bool:FEATURE_BO_SIGNALEMENT_CREATE)%'
     send_error_email_token: '%env(SEND_ERROR_EMAIL_TOKEN)%'
-    feature_sites_faciles: '%env(bool:FEATURE_SITES_FACILES)%'
     feature_banner_histologe: '%env(bool:FEATURE_BANNER_HISTOLOGE)%'
     feature_secure_uuid_url: '%env(bool:FEATURE_SECURE_UUID_URL)%'
