+++ src/Controller/HomepageController.php
@@ -3,36 +3,25 @@
 namespace App\Controller;
 
 use App\Dto\DemandeLienSignalement;
-use App\Form\ContactType;
 use App\Form\DemandeLienSignalementType;
-use App\Form\PostalCodeSearchType;
-use App\FormHandler\ContactFormHandler;
 use App\Repository\SignalementRepository;
 use App\Service\Mailer\NotificationMail;
 use App\Service\Mailer\NotificationMailerRegistry;
 use App\Service\Mailer\NotificationMailerType;
-use App\Service\Signalement\PostalCodeHomeChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
-use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\Cache;
 use Symfony\Component\RateLimiter\RateLimiterFactory;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Contracts\Cache\CacheInterface;
-use Symfony\Contracts\Cache\ItemInterface;
 
 class HomepageController extends AbstractController
 {
     public function __construct(
-        private readonly CacheInterface $cache,
         #[Autowire(env: 'SITES_FACILES_URL')]
         private readonly string $sitesFacilesUrl,
-        #[Autowire(env: 'FEATURE_SITES_FACILES')]
-        private readonly bool $featureSitesFaciles,
     ) {
     }
 
@@ -42,61 +31,14 @@ class HomepageController extends AbstractController
         defaults: ['show_sitemap' => true]
     )]
     public function index(
-        Request $request,
-        SignalementRepository $signalementRepository,
-        PostalCodeHomeChecker $postalCodeHomeChecker,
         #[Autowire(param: 'kernel.environment')]
         string $environment,
     ): Response {
-        if ($this->featureSitesFaciles && 'prod' === $environment) {
+        if ('prod' === $environment) {
             return $this->redirect($this->sitesFacilesUrl, Response::HTTP_MOVED_PERMANENTLY);
         }
-        if ($this->featureSitesFaciles) {
-            return $this->redirectToRoute('app_login');
-        }
 
-        $stats = $this->cache->get('homepage.stats.array', function (ItemInterface $item) use ($signalementRepository) {
-            $item->expiresAfter(900);
-
-            $stats = ['pris_en_compte' => 0, 'clotures' => 0];
-            $stats['total'] = $signalementRepository->countAll(
-                territory: null,
-                partners: null,
-                removeImported: true,
-                removeArchived: true
-            );
-
-            if ($stats['total'] > 0) {
-                $stats['pris_en_compte'] = round($signalementRepository->countValidated(true) / $stats['total'] * 100, 1);
-                $stats['clotures'] = round($signalementRepository->countClosed(true) / $stats['total'] * 100, 1);
-            }
-
-            return $stats;
-        });
-
-        $displayModal = '';
-
-        $form = $this->createForm(PostalCodeSearchType::class, []);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $inputPostalCode = $form->get('postalcode')->getData();
-            if ($postalCodeHomeChecker->isActiveByPostalCode($inputPostalCode)) {
-                return $this->redirectToRoute('front_signalement', ['cp' => $inputPostalCode]);
-            }
-            $displayModal = $inputPostalCode;
-        }
-
-        $demandeLienSignalement = new DemandeLienSignalement();
-        $formDemandeLienSignalement = $this->createForm(DemandeLienSignalementType::class, $demandeLienSignalement, [
-            'action' => $this->generateUrl('front_demande_lien_signalement'),
-        ]);
-
-        return $this->render('front/index.html.twig', [
-            'form_postalcode' => $form->createView(),
-            'stats' => $stats,
-            'display_modal' => $displayModal,
-            'formDemandeLienSignalement' => $formDemandeLienSignalement,
-        ]);
+        return $this->redirectToRoute('app_login');
     }
 
     #[Route('/demande-lien-signalement', name: 'front_demande_lien_signalement', methods: ['POST'])]
@@ -151,11 +93,7 @@ class HomepageController extends AbstractController
     )]
     public function about(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'a-propos/qui-sommes-nous/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/about.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'a-propos/qui-sommes-nous/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -165,13 +103,7 @@ class HomepageController extends AbstractController
     )]
     public function obligations_entretien(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'blog/entretien-logement-qui-paye-quoi/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/obligations_entretien.html.twig', [
-            'guide_path' => 'build/files/GUIDE-QUI-REPARE-QUI-ENTRETIENT-Ministere2016-1.pdf',
-        ]);
+        return $this->redirect($this->sitesFacilesUrl.'blog/entretien-logement-qui-paye-quoi/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -181,11 +113,7 @@ class HomepageController extends AbstractController
     )]
     public function aides_travaux(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'blog/quelles-aides-pour-faire-des-travaux-dans-mon-logement/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/aides_travaux.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'blog/quelles-aides-pour-faire-des-travaux-dans-mon-logement/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -194,63 +122,9 @@ class HomepageController extends AbstractController
         methods: ['GET'],
         defaults: ['show_sitemap' => false]
     )]
-    public function contact(
-        ParameterBagInterface $parameterBag,
-    ): Response {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'une-question/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-        $form = $this->createForm(ContactType::class, null, [
-            'action' => $this->generateUrl('front_contact_form'),
-        ]);
-        $demandeLienSignalement = new DemandeLienSignalement();
-        $formDemandeLienSignalement = $this->createForm(DemandeLienSignalementType::class, $demandeLienSignalement, [
-            'action' => $this->generateUrl('front_demande_lien_signalement'),
-        ]);
-
-        return $this->render('front/contact.html.twig', [
-            'form' => $form->createView(),
-            'contactEmail' => $parameterBag->get('contact_email'),
-            'formDemandeLienSignalement' => $formDemandeLienSignalement,
-        ]);
-    }
-
-    #[Route('/contact', name: 'front_contact_form', methods: ['POST'])]
-    public function contactForm(
-        Request $request,
-        ContactFormHandler $contactFormHandler,
-        RateLimiterFactory $contactFormLimiter,
-    ): Response {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'une-question/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-        $type = 'error';
-        $message = null;
-        $form = $this->createForm(ContactType::class, null, [
-            'action' => $this->generateUrl('front_contact_form'),
-        ]);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $limiter = $contactFormLimiter->create($request->getClientIp());
-            if (false === $limiter->consume(1)->isAccepted()) {
-                $message = 'Vous avez atteint le nombre maximum de messages que vous pouvez envoyer. Veuillez réessayer plus tard.';
-                $view = $this->renderView('form/form-contact.html.twig', ['form' => $form, 'type' => $type, 'message' => $message]);
-
-                return new JsonResponse(['html' => $view]);
-            }
-            $contactFormHandler->handle(
-                $form->get('nom')->getData(),
-                $form->get('email')->getData(),
-                $form->get('message')->getData(),
-                (string) $form->get('organisme')->getData(),
-                $form->get('objet')->getData()
-            );
-            $type = 'success';
-            $message = 'Votre message à bien été envoyé !';
-        }
-        $view = $this->renderView('form/form-contact.html.twig', ['form' => $form, 'type' => $type, 'message' => $message]);
-
-        return new JsonResponse(['html' => $view]);
+    public function contact(): Response
+    {
+        return $this->redirect($this->sitesFacilesUrl.'une-question/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -260,11 +134,7 @@ class HomepageController extends AbstractController
     )]
     public function cguUsager(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'cgu', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/cgu_usagers.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'cgu', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -274,11 +144,7 @@ class HomepageController extends AbstractController
     )]
     public function cguPro(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'cgu-agents/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/cgu_agents.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'cgu-agents/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -288,11 +154,7 @@ class HomepageController extends AbstractController
     )]
     public function politiqueConfidentialite(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'politique-de-confidentialite/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/politique_de_confidentialite.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'politique-de-confidentialite/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -302,11 +164,7 @@ class HomepageController extends AbstractController
     )]
     public function mentionsLegales(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'mentions-legales/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/mentions_legales.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'mentions-legales/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route(
@@ -316,45 +174,19 @@ class HomepageController extends AbstractController
     )]
     public function accessibilite(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl.'accessibilite/', Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/accessibilite.html.twig');
+        return $this->redirect($this->sitesFacilesUrl.'accessibilite/', Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Route('/plan-du-site', name: 'plan_du_site')]
     public function planDuSite(): Response
     {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl, Response::HTTP_MOVED_PERMANENTLY);
-        }
-
-        return $this->render('front/plan_du_site.html.twig');
+        return $this->redirect($this->sitesFacilesUrl, Response::HTTP_MOVED_PERMANENTLY);
     }
 
     #[Cache(public: true, maxage: 3600)]
     #[Route('/sitemap.{_format}', name: 'app_front_sitemap', defaults: ['_format' => 'xml'])]
-    public function generateSitemap(
-        RouterInterface $router,
-        #[Autowire(param: 'host_url')]
-        string $hostUrl,
-    ) {
-        if ($this->featureSitesFaciles) {
-            return $this->redirect($this->sitesFacilesUrl, Response::HTTP_MOVED_PERMANENTLY);
-        }
-        $urls = [];
-        $routes = $router->getRouteCollection()->all();
-        foreach ($routes as $route) {
-            if ($route->getDefaults()['show_sitemap'] ?? false) {
-                $urls[] = ['loc' => $hostUrl.$route->getPath()];
-            }
-        }
-
-        return new Response(
-            $this->renderView('front/sitemap.xml.twig', ['urls' => $urls]),
-            Response::HTTP_OK,
-            ['Content-Type' => 'text/xml']
-        );
+    public function generateSitemap()
+    {
+        return $this->redirect($this->sitesFacilesUrl, Response::HTTP_MOVED_PERMANENTLY);
     }
 }
