+++ templates/back/nav_bo.html.twig
@@ -1,5 +1,5 @@
 {% set notification_count = count_notification(app.user) %}
-{% set logoClass = platform.name == 'histologe' ? 'logo-width' : 'logo-width-lg' %}
+{% set logoClass = 'logo-width-lg' %}
 <header class="fr-header">
     <div class="fr-header__body">
         <div class="{{ container_class }}">
