--- templates/front/contact.html.twig
@@ -1,60 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Contacter{% endblock %}
-
-{% block body %}
-    {% include "breadcrumbs.html.twig" with { 'breadcrumbsLabel': 'Contact' } %}
-	<main id="content">
-		<section class="fr-container">
-			<h1 class="fr-mt-5w">Contacter {{ platform.name }}</h1>
-		</section>
-		<section class="fr-container">
-			<h2>Je souhaite signaler un problème de logement</h2>
-			<p>
-				Vous êtes au bon endroit ! {{ platform.name }} vous accompagne dans le signalement de vos problèmes de logement. 
-				Cliquez sur le bouton ci-dessous pour déposer votre signalement.
-			</p>
-			<a href="{{ path('front_signalement') }}" class="fr-btn fr-btn--icon-left fr-arrow-up-line">Je dépose un signalement</a>
-		</section>
-		<section class="fr-container">
-			{% include '_partials/_demande-lien-signalement-intro.html.twig' %}
-			{% include '_partials/_demande-lien-signalement-container.html.twig' %}
-		</section>
-		<section class="fr-container">
-			<h2 class="fr-mt-5w">J'ai une question sur {{ platform.name }}</h2>
-			<p>
-				Vous trouverez peut-être une réponse dans les ressources suivantes :
-			</p>
-			<ul>
-				<li><a href="{{ gitbook.faq }}" target="_blank" rel="noopener" title="La documentation {{ platform.name }} - ouvre une nouvelle fenêtre">La documentation {{ platform.name }}</a></li>
-				<li><a href="{{ gitbook.faq }}/usager/signalement/jai-depose-un-signalement-et-maintenant" target="_blank" rel="noopener" title="documentation {{ platform.name }} : J'ai déposé un signalement, et maintenant ? - ouvre une nouvelle fenêtre">J'ai déposé un signalement, et maintenant ?</a></li>
-				<li><a href="{{path('front_cgu')}}">Conditions générales d'utilisation</a></li>
-				<li><a href="{{path('politique_de_confidentialite')}}">Politique de confidentialité</a></li>
-			</ul>
-		</section>
-		<section class="fr-container">
-			<h2 class="fr-mt-5w">Vous avez une autre question ?</h2>
-			<p>Vous ne trouvez pas la réponse à votre question ? Vous souhaitez signaler un problème technique ? Contactez l'équipe {{ platform.name }}.</p>
-			<div class="fr-alert fr-alert--warning fr-my-3w">
-				<h3 class="fr-alert__title">N'envoyez pas votre signalement par e-mail !</h3>
-				<p>
-					Les signalements ne peuvent être pas déposés par e-mail. 
-					Pour déposer un signalement, cliquez sur le bouton "Je dépose mon signalement" sur cette page ou sur l'élément "Signalement" dans le menu principal.
-				</p>
-			</div>
-			<a href="mailto:{{ contactEmail }}" class="fr-btn fr-btn--icon-left fr-mail-line">Envoyer un message</a>
-		</section>
-		<section class="fr-py-5w fr-mt-5w fr-background-contrast--blue-france">
-			<div class="fr-container">
-				<h2>Vous êtes acteur de la lutte contre l'habitat indigne et souhaitez en savoir plus sur {{ platform.name }} ?</h2>
-				<p>
-					Administrations, départements, associations, bailleurs… {{ platform.name }} rassemble les acteurs de la lutte contre l'habitat indigne. 
-					Si vous souhaitez en savoir plus sur notre service, n'hésitez pas à nous contacter.
-				</p>
-				<div id="container-form-contact">
-					{% include 'form/form-contact.html.twig' with {'form': form} only %}
-				</div>
-			</div>
-		</section>
-	</main>
-{% endblock %}
