--- templates/front/mentions_legales.html.twig
@@ -1,73 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}Mentions légales {% endblock %}
-
-{% block body %}
-    {% include "breadcrumbs.html.twig" with { 'breadcrumbsLabel': 'Mentions légales' } %}
-
-    <main class="fr-container fr-py-1w" id="content">
-        <section class="cgu fr-grid-row">
-            <div class="fr-col-12 fr-col-lg-8">
-                <h1>Mentions légales de la Plateforme {{ platform.name }}</h1>
-
-                <h2>Editeur de la Plateforme</h2>
-
-                <p>La Plateforme « {{ platform.name }} » est éditée par :</p>
-                <p>La Direction générale de l’aménagement, du logement et de la nature (DGALN) située :</p>
-
-                <address>
-                    Tour Séquoia<br>
-                    1 place Carpeaux<br>
-                    92800 Puteaux<br>
-                    France<br><br>
-                    Téléphone : <a href="tel:+33140812122">01 40 81 21 22</a> <br><br>
-                </address>
-
-                <h2>Directeur de la publication</h2>
-                <p>
-                    Le Directeur de la publication est Monsieur Philippe MAZENC, directeur général de l’aménagement, du
-                    logement et de la nature.
-                </p>
-
-                <h2>Hébergement de la Plateforme</h2>
-
-                <p>Cette Plateforme est hébergée par :</p>
-                <address>
-                    Scalingo SAS<br>
-                    13 rue Jacques Peirotes<br>
-                    67000 Strasbourg<br>
-                    France<br><br>
-                </address>
-
-                <h2>Accessibilité</h2>
-                <p>
-                    La conformité aux normes d’accessibilité numérique est un objectif ultérieur mais nous tâchons de
-                    rendre cette plateforme accessible à toutes et à tous.
-                </p>
-                <h2>En savoir plus</h2>
-                <p>
-                    Pour en savoir plus sur la politique d’accessibilité numérique de l’État : <a
-                            href="https://accessibilite.numerique.gouv.fr/" target="_blank" rel="noreferrer noopener">https://accessibilite.numerique.gouv.fr/</a>
-                </p>
-                <h2>Signaler un dysfonctionnement</h2>
-                <p>
-                    Si vous rencontrez un défaut d’accessibilité vous empêchant d’accéder à un contenu ou une
-                    fonctionnalité de la plateforme, merci de nous en faire part ici : <a
-                            href="{{ path('front_contact') }}"> {{ platform.url ~ '/contact' }}</a>
-                </p>
-                <p>
-                    Si vous n’obtenez pas de réponse rapide de notre part, vous êtes en droit de faire parvenir vos
-                    doléances ou une demande de saisine au Défenseur des droits.
-                </p>
-                <h2>Sécurité</h2>
-                <p>
-                    La plateforme est protégée par un certificat électronique, matérialisé pour la grande majorité des
-                    navigateurs par un cadenas. Cette protection participe à la confidentialité des échanges.
-                </p>
-                <p>
-                    En aucun cas les services associés à la plateforme ne seront à l’origine d’envoi de courriels pour
-                    demander la saisie d’informations personnelles.
-                </p>
-            </div>
-        </section>
-    </main>
-{% endblock %}
