--- templates/front/plan_du_site.html.twig
@@ -1,109 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Plan du site{% endblock %}
-
-{% block body %}
-    <div class="fr-container fr-py-16v plan-du-site-header" id="content">
-        <div class="fr-grid-row">
-            <div class="fr-col-12 fr-col-md-8">
-                <h1>Plan du site</h1>
-                <section class="fr-py-6v">
-                    <h2>Accueil et pages génériques</h2>
-                    <ul>
-                        <li>
-                            <a href="{{ path('home') }}">Accueil</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_about') }}">Qui sommes-nous ?</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_contact') }}">Contact</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_signalement') }}">Signaler un problème de logement</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_statistiques') }}">Statistiques d'usage</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_obligations_entretien') }}">Entretien du logement : les obligations</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_aides_travaux') }}">Les aides aux travaux</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_cgu') }}">Conditions générales d'utilisation</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_cgu_agents') }}">Conditions générales d'utilisation de l'espace pro</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('mentions_legales') }}">Mentions légales</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('politique_de_confidentialite') }}">Politique de confidentialité</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('front_accessibilite') }}">Déclaration d'accessibilité</a>
-                        </li> 
-                    </ul>
-                </section>
-                <section class="fr-py-6v">
-                    <h2>Gestion du compte</h2>
-                    <ul>
-                        <li>
-                            <a href="{{ path('app_login') }}">Se connecter à l'espace pro</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('login_activation') }}">Activer mon compte</a>
-                        </li>
-                        <li>
-                            <a href="{{ path('login_mdp_perdu') }}">Réinitialiser mon mot de passe</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('app_logout') }}">Déconnexion du compte (connexion nécessaire)</a>
-                        </li>
-                    </ul>
-                </section>
-                <section class="fr-py-6v">
-                    <h2>Tableau de bord utilisateur</h2>
-                    <ul>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_profil') }}">Profil utilisateur (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_dashboard') }}">Tableau de bord utilisateur (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_signalements_index') }}">Liste des signalements (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_notifications_list') }}">Notifications (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_cartographie') }}">Cartographie (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_statistiques') }}">Statistiques (connexion nécessaire)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_partner_index') }}">Liste des partenaires (connexion nécessaire, avec droits particuliers)</a>
-                        </li>
-                        <li>
-                            <span class="fr-icon-account-line" title="Page réservée aux utilisateurs connectés" aria-hidden="true"></span>
-                            <a href="{{ path('back_tags_index') }}">Liste des étiquettes (connexion nécessaire, avec droits particuliers)</a>
-                        </li>
-                    </ul>
-                </section>
-            </div>
-        </div>
-    </div>
-{% endblock %}
