--- templates/front/sitemap.xml.twig
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-    {% for url in urls %}
-        <url>
-            <loc>{{ url.loc }}</loc>
-        </url>
-    {% endfor %}
-</urlset>
